:art: move websocket processing logic to services module
